scripts/mflt-build-id/pyproject.toml
[tool.poetry]
name = "mflt-build-id"
version = "0.0.5"
description = "Memfault Build Id injector"
authors = ["Memfault Inc <hello@memfault.com>"]
repository = "https://github.com/memfault/memfault-firmware-sdk"
packages = [
  { include = "mflt_build_id", from = "src" }
]

[tool.poetry.scripts]
mflt_build_id = 'mflt_build_id:main'

[tool.poetry.dependencies]
python = "^3.6" # must match memfault-cli
pyelftools = ">=0.26,<=0.28"

[tool.poetry.dev-dependencies]
pytest = "^6"
pytest-cov = "^2"
pytest-timeout = "^2.0"
wasmer = "0.3.0"

[tool.black]
line-length = 100
target-version = ['py36']
include = '\.pyi?$'

[tool.isort]
# Black Compatible Settings from
# https://black.readthedocs.io/en/stable/the_black_code_style.html?highlight=isort#how-black-wraps-lines
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

#
# NB: Uncommenting these lines will prevent installing the project in editable mode.
#
# [build-system]
# requires = ["poetry>=0.12"]
# build-backend = "poetry.masonry.api"
